requirements.txt
fastapi==0.103.1
pydantic==2.3.0
python-dotenv==1.0.0
python_dateutil==2.8.2
requests==2.31.0
